refactor(ttnn): split wavelet device operations
