+++ ttnn/cpp/ttnn/operations/wavelet/CMakeLists.txt
@@ -40,12 +40,12 @@ target_sources(
         ${TTNN_OP_WAVELET_SRCS}
 )
 
-# The two factories deliberately keep independent planner helpers and each
-# instantiates all 106 static schemes. Compile them outside the unity source to
-# avoid anonymous-namespace collisions and an excessively large translation unit.
+# The implementation sources instantiate all 106 static schemes. Compile them
+# outside the unity source to avoid anonymous-namespace collisions and an
+# excessively large translation unit.
 set_source_files_properties(
-    device/wavelet_1d_program_factory.cpp
-    device/wavelet_2d_program_factory.cpp
+    device/wavelet_1d_operation_impl.cpp
+    device/wavelet_2d_operation_impl.cpp
     PROPERTIES
         SKIP_UNITY_BUILD_INCLUSION
             ON
