+++ ttnn/cpp/ttnn/operations/wavelet/sources.cmake
@@ -5,8 +5,16 @@ set(TTNN_OP_WAVELET_API_HEADERS
 
 set(TTNN_OP_WAVELET_SRCS
     common/wavelet_host.cpp
-    device/wavelet_1d_program_factory.cpp
-    device/wavelet_2d_program_factory.cpp
+    device/ilwt_1d_device_operation.cpp
+    device/ilwt_1d_program_factory.cpp
+    device/ilwt_2d_device_operation.cpp
+    device/ilwt_2d_program_factory.cpp
+    device/lwt_1d_device_operation.cpp
+    device/lwt_1d_program_factory.cpp
+    device/lwt_2d_device_operation.cpp
+    device/lwt_2d_program_factory.cpp
+    device/wavelet_1d_operation_impl.cpp
+    device/wavelet_2d_operation_impl.cpp
     wavelet.cpp
 )
